feat(settings): add commands to insert em-dash and en-dash symbols
Due to me using alt+- and alt+shift+- as undo/redo, macOS's which are macOS' default keybindings for typing en-dash and em-dash, I need an alternative way to insert them when needed.

This solution uses a combo of the macros and aliases extensions to achieve it. It's a bit of a dirty hack, but works for now. At a later day I'll look for a solution that lets me define and add custom commands.